+++ requests_cache/models/request.py
@@ -16,17 +16,29 @@ class CachedRequest:
     """A serializable dataclass that emulates :py:class:`requests.PreparedResponse`"""
 
     body: bytes = field(default=None, converter=encode)
-    cookies: RequestsCookieJar = field(factory=dict)
+    cookies: RequestsCookieJar = field(factory=RequestsCookieJar)
     headers: CaseInsensitiveDict = field(factory=CaseInsensitiveDict)
     method: str = field(default=None)
     url: str = field(default=None)
 
     @classmethod
-    def from_request(cls, original_request: PreparedRequest):
+    def from_request(cls, original_request: PreparedRequest) -> 'CachedRequest':
         """Create a CachedRequest based on an original request object"""
         kwargs = {k: getattr(original_request, k, None) for k in fields_dict(cls).keys()}
-        kwargs['cookies'] = original_request._cookies
+        kwargs['cookies'] = getattr(original_request, '_cookies', None)
         return cls(**kwargs)
+
+    def prepare(self) -> PreparedRequest:
+        """Convert the CachedRequest back into a PreparedRequest"""
+        prepared_request = PreparedRequest()
+        prepared_request.prepare(
+            cookies=self.cookies,
+            data=self.body,
+            headers=self.headers,
+            method=self.method,
+            url=self.url,
+        )
+        return prepared_request
 
     @property
     def _cookies(self):