+++ tests/conftest.py
@@ -13,6 +13,7 @@ import pytest
 from datetime import datetime, timezone
 from functools import wraps
 from logging import basicConfig, getLogger
+from os.path import abspath, dirname, join
 from tempfile import NamedTemporaryFile
 
 import requests
@@ -57,6 +58,10 @@ MOCKED_URL_REDIRECT = 'http+mock://requests-cache.com/redirect'
 MOCKED_URL_REDIRECT_TARGET = 'http+mock://requests-cache.com/redirect_target'
 MOCKED_URL_404 = 'http+mock://requests-cache.com/nonexistent'
 MOCK_PROTOCOLS = ['mock://', 'http+mock://', 'https+mock://']
+
+PROJECT_DIR = abspath(dirname(dirname(__file__)))
+SAMPLE_DATA_DIR = join(PROJECT_DIR, 'tests', 'sample_data')
+SAMPLE_CACHE_FILES = [join(SAMPLE_DATA_DIR, path) for path in os.listdir(SAMPLE_DATA_DIR)]
 
 AWS_OPTIONS = {
     'endpoint_url': 'http://localhost:8000',
@@ -116,7 +121,6 @@ def mock_session() -> CachedSession:
             db_path=temp.name,
             backend='sqlite',
             allowable_methods=ALL_METHODS,
-            suppress_warnings=True,
         )
         yield mount_mock_adapter(session)
 
@@ -129,7 +133,6 @@ def tempfile_session() -> CachedSession:
             cache_name=temp.name,
             backend='sqlite',
             allowable_methods=ALL_METHODS,
-            suppress_warnings=True,
         )
         yield session
 
@@ -151,7 +154,6 @@ def installed_session() -> CachedSession:
             cache_name=temp.name,
             backend='sqlite',
             allowable_methods=ALL_METHODS,
-            suppress_warnings=True,
         )
         yield requests.Session()
     requests_cache.uninstall_cache()