+++ HISTORY.md
@@ -1,6 +1,13 @@
 # History
 
-## 0.7.0 (2021-07-TBD)
+## 0.7.2 (2021-07-TBD)
+* Add support for `Response.next` when 302 responses are cached directly
+* Make `CachedResponse` a non-slotted class to allow client code to set arbitrary attributes on it
+
+## 0.7.1 (2021-07-09)
+* Fix a bug in which Cache-Control headers would be used unexpectedly
+
+## 0.7.0 (2021-07-07)
 [See all issues and PRs for 0.7](https://github.com/reclosedev/requests-cache/milestone/2?closed=1)
 
 ### Backends
@@ -39,6 +46,7 @@
 ### General
 * Add option to manually cache response objects with `BaseCache.save_response()`
 * Add `BaseCache.keys()` and `values()` methods
+* Add `BaseCache.response_count()` method to get an accurate count of responses (excluding invalid and expired)
 * Show summarized response details with `str(CachedResponse)`
 * Add more detailed repr methods for `CachedSession`, `CachedResponse`, and `BaseCache`
 * Add support for caching multipart form uploads
@@ -59,6 +67,7 @@
     * `requests-cache[redis]`
 * Packaging is now handled with Poetry. For users, installation still works the same. For developers,
   see [Contributing Guide](https://requests-cache.readthedocs.io/en/stable/contributing.html) for details
+* requests-cache is now fully typed and PEP-561 compliant
 
 -----
 ### 0.6.4 (2021-06-04)
@@ -112,7 +121,7 @@ next time they are requested. They can also be manually converted or removed, if
 * Add [example script](https://github.com/reclosedev/requests-cache/blob/master/examples/convert_cache.py)
   to convert an existing cache from previous serialization format to new one
 * When running `remove_expired_responses()`, also remove responses that are invalid due to updated
-  serialization format 
+  serialization format
 * Add `CachedResponse` class to wrap cached `requests.Response` objects, which makes additional
   cache information available to client code
 * Add `CachedHTTPResponse` class to wrap `urllib3.response.HTTPResponse` objects, available via `CachedResponse.raw`